Se agrega estado de partida y contado de espera antes la ola de enemigos

+++ Assets/Game/Scripts/ScriptableObjects/WaveProfile.cs
@@ -8,6 +8,6 @@ namespace TowerDefender.Game.ScriptableObjects
     {
         public int EnemyCount;
         public float SpawnFrequency;
-        public float CoolDownTime;
+        public int CoolDownTime;
     }
 }
